PowerVR Graphics SDK v5.12 Release.
SDK 24.1(v5.12) release.

Native SDK:
* Changes in CMake to speed up Android builds. Downgraded CMake to v3.10 and Gradle to v3.5.3. NDK 20.0.5594570 is a valid version when cross compiling for Android from Windows / Linux.
* Added new Vulkan example: Antialiasing (MSAA, FXAA, TAA).
* Fixed VulkanTimelineSemaphores example crashing on early exit when the extension is not supported.
* Fixed VulkanSubgroups example no ouput on BXM-8-256: Workgroup width was being set to zero during setup.
* Fixed OpenGLESPostProcessing and VulkanPostProcessing small artifacts on the edge of the mesh when no post processing method is being applied.
* Fix: Making several SDK demos time deterministic when using the -forceframetime parameter (OpenGLESDeferredShading, OpenGLESGameOfLife, OpenGLESParticleSystem, VulkanDeferredShading, VulkanDeferredShadingPFX, VulkanGameOfLife, VulkanGnomeHorde, ParticleSystemGPU).
* Fixed VulkanHelloRayTracing various buffers being built with not optimal memory properties.
* Fixed PVRVk getCacheMaxDataSize(), initializing dataSize parameter handed to vkGetPipelineCacheData to identify cases where pipeline cached data has size 0.
* Fixed over 170 compilation warnings.

PVRVFrame:
* Fix the OpenGL ES 2+ implementation of 'glBlendEquation' to correctly handle the advanced blend equation modes.
* Fix libGLESv2.so dynamic loading error symbol lookup error: undefined symbol which prevented successful loading of libGLESv2.so in some code paths.
* Add support for GL_EXT_sparse_texture on platforms that support GL_ARB_sparse_texture.
* Add support for GL_EXT_clip_control on platforms that support GL_ARB_clip_control or OpenGL 4.5.
* Fix glAttachShader so it returns GL_INVALID_OPERATION if a shader is passed in of a type that is already attached to the target program.
* Fix glReadPixels reading from framebuffers that make use of glFramebufferTexture2DMultisampleEXT|IMG.

The documentation from this repository has been moved to https://docs.imgtec.com/

Refer to our release notes for more details about the changes: https://developer.imaginationtech.com/tools/release-notes/

+++ examples/Vulkan/04_IntroducingPVRUtils/VulkanIntroducingPVRUtils.cpp
@@ -230,10 +230,10 @@ pvr::Result VulkanIntroducingPVRUtils::initView()
 	_deviceResources->commandPool = _deviceResources->device->createCommandPool(pvrvk::CommandPoolCreateInfo(queueAccessInfo.familyId));
 
 	_deviceResources->descriptorPool = _deviceResources->device->createDescriptorPool(pvrvk::DescriptorPoolCreateInfo()
-																						  .addDescriptorInfo(pvrvk::DescriptorType::e_COMBINED_IMAGE_SAMPLER, 8 * swapchainLength)
-																						  .addDescriptorInfo(pvrvk::DescriptorType::e_UNIFORM_BUFFER_DYNAMIC, 8 * swapchainLength)
-																						  .addDescriptorInfo(pvrvk::DescriptorType::e_UNIFORM_BUFFER, 8 * swapchainLength)
-																						  .setMaxDescriptorSets(8 * swapchainLength));
+																						  .addDescriptorInfo(pvrvk::DescriptorType::e_COMBINED_IMAGE_SAMPLER, static_cast<uint16_t>(8 * swapchainLength))
+																						  .addDescriptorInfo(pvrvk::DescriptorType::e_UNIFORM_BUFFER_DYNAMIC, static_cast<uint16_t>(8 * swapchainLength))
+																						  .addDescriptorInfo(pvrvk::DescriptorType::e_UNIFORM_BUFFER, static_cast<uint16_t>(8 * swapchainLength))
+																						  .setMaxDescriptorSets(static_cast<uint16_t>(8 * swapchainLength)));
 
 	// create demo buffers
 	createBuffers();
