+++ examples/Vulkan/GaussianBlur/VulkanGaussianBlur.cpp
@@ -558,7 +558,7 @@ pvr::Result VulkanGaussianBlur::initView()
 	_deviceResources->commandPool = _deviceResources->device->createCommandPool(pvrvk::CommandPoolCreateInfo(_deviceResources->queues[0]->getFamilyIndex()));
 
 	_deviceResources->descriptorPool = _deviceResources->device->createDescriptorPool(
-		pvrvk::DescriptorPoolCreateInfo(8 * _swapchainLength).addDescriptorInfo(pvrvk::DescriptorType::e_STORAGE_IMAGE, 8 * _swapchainLength));
+		pvrvk::DescriptorPoolCreateInfo(static_cast<uint16_t>(8 * _swapchainLength)).addDescriptorInfo(pvrvk::DescriptorType::e_STORAGE_IMAGE, static_cast<uint16_t>(8 * _swapchainLength)));
 
 	// Create per frame resource
 	for (uint32_t i = 0; i < _swapchainLength; ++i)
