+++ examples/Vulkan/Glass/VulkanGlass.cpp
@@ -1363,10 +1363,10 @@ pvr::Result VulkanGlass::initView()
 	//---------------
 	// Create the DescriptorPool
 	pvrvk::DescriptorPoolCreateInfo descPoolInfo;
-	descPoolInfo.addDescriptorInfo(pvrvk::DescriptorType::e_COMBINED_IMAGE_SAMPLER, 12 * _swapchainLength)
-		.addDescriptorInfo(pvrvk::DescriptorType::e_UNIFORM_BUFFER_DYNAMIC, 12 * _swapchainLength)
-		.addDescriptorInfo(pvrvk::DescriptorType::e_UNIFORM_BUFFER, 12 * _swapchainLength)
-		.setMaxDescriptorSets(12 * _swapchainLength);
+	descPoolInfo.addDescriptorInfo(pvrvk::DescriptorType::e_COMBINED_IMAGE_SAMPLER, static_cast<uint16_t>(12 * _swapchainLength))
+		.addDescriptorInfo(pvrvk::DescriptorType::e_UNIFORM_BUFFER_DYNAMIC, static_cast<uint16_t>(12 * _swapchainLength))
+		.addDescriptorInfo(pvrvk::DescriptorType::e_UNIFORM_BUFFER, static_cast<uint16_t>(12 * _swapchainLength))
+		.setMaxDescriptorSets(static_cast<uint16_t>(12 * _swapchainLength));
 
 	_deviceResources->descriptorPool = _deviceResources->device->createDescriptorPool(descPoolInfo);
